Fixed Dart "firstWhere orElse null" Issue
Fixed issue that the orElse property of firstWhere method of Iterables can't be null if any methods inside the Individual iterables are accessed.
Had to replace the firstWhere method with for loops. Can't complain though. There is nothing better than "for" loops.
Bumped Version and Updated Dependencies of the package.